+++ exts/events/events.py
@@ -30,10 +30,11 @@ if TYPE_CHECKING:
     from core.bot import ziBot
 
 
-REASON_REGEX = re.compile(r"^\[\S+\#\d+ \(ID: (\d+)\)\]: (.*)")
+REASON_REGEX = re.compile(
+    r"^\[\S+\#\d+ \(ID: (?P<userId>[0-9]+)\) #(?P<caseNum>[0-9]+)\]: (?P<reason>.*)"
+)
 
 
-# TODO: Move this to exts.utils.other
 async def doModlog(
     bot: ziBot,
     guild: discord.Guild,
@@ -56,9 +57,10 @@ async def doModlog(
             # Get the real moderator
             match = REASON_REGEX.match(reason)
             if match:
-                modId = int(match.group(1))
+                modId = int(match.group("userId"))
                 moderator = bot.get_user(modId) or await bot.fetch_user(modId)
-                reason = match.group(2)
+                reason = match.group("reason")
+                caseNum = int(match.group("caseNum"))
 
     else:
         # Since moderation is done manually, caselog will be done here
@@ -75,17 +77,18 @@ async def doModlog(
         # No channel found, don't do modlog
         return
 
+    if not caseNum:
+        # placeholder for failed case parsing attempt
+        caseNum = -1
+
     e = ZEmbed.minimal(
-        title="Modlog - {}".format(type.title()),
+        title="{} | #{}".format(type.title(), caseNum),
         description=(
             f"**User**: {member} ({member.mention})\n"
             + (f"**Reason**: {reason}\n" if reason else "")
             + f"**Moderator**: {moderator.mention}"
         ),
     )
-    # TODO: Make caseNum appears when moderation cmd is used
-    if caseNum is not None:
-        e.description += f"\n**Case**: #{caseNum}"
 
     e.set_footer(text=f"ID: {member.id}")
     await channel.send(embed=e)  # type: ignore
@@ -165,18 +168,23 @@ class EventHandler(commands.Cog, CogMixin):
                 # Something wrong happened
                 return
 
+    async def getAuditLogs(
+        self, guild: discord.Guild, limit=1, delay=2, **kwargs
+    ) -> discord.AuditLogEntry:
+        # discord needs a few second to update Audit Logs
+        await asyncio.sleep(delay)
+        return (await guild.audit_logs(limit=limit, **kwargs).flatten())[0]
+
     @commands.Cog.listener("on_member_remove")
     async def onMemberRemove(self, member: discord.Member) -> None:
         """Farewell message"""
         guild: discord.Guild = member.guild
 
         with suppress(discord.Forbidden):
-            # discord needs a few second to update Audit Logs
-            await asyncio.sleep(2)
-
-            entry = (await guild.audit_logs(limit=1).flatten())[0]
+            entry = await self.getAuditLogs(guild)
 
             if entry.target == member:
+
                 # TODO: Filters bot's action
                 if entry.action == discord.AuditLogAction.kick:
                     self.bot.dispatch("member_kick", member, entry)
@@ -185,6 +193,11 @@ class EventHandler(commands.Cog, CogMixin):
                 if entry.action == discord.AuditLogAction.ban:
                     # Intents.bans are disabled to make this works
                     self.bot.dispatch("member_ban", member, entry)
+                    return
+
+                if entry.action == discord.AuditLogAction.unban:
+                    # Intents.bans are disabled to make this works
+                    self.bot.dispatch("member_unban", member, entry)
                     return
 
         # fallback to farewell message
@@ -204,17 +217,30 @@ class EventHandler(commands.Cog, CogMixin):
         )
 
     @commands.Cog.listener("on_member_ban")
-    async def onMemberBan(
-        self, member: discord.Member, entry: discord.AuditLogEntry
-    ) -> None:
-        await doModlog(
-            self.bot,
-            member.guild,
-            entry.target,  # type: ignore
-            entry.user,
-            "ban",
-            entry.reason,
-        )
+    async def onMemberBan(self, guild: discord.Guild, user: discord.User) -> None:
+        entry = await self.getAuditLogs(guild)
+        if entry.target == user:
+            await doModlog(
+                self.bot,
+                guild,
+                entry.target,  # type: ignore
+                entry.user,
+                "ban",
+                entry.reason,
+            )
+
+    @commands.Cog.listener("on_member_unban")
+    async def onMemberUnban(self, guild: discord.Guild, user: discord.User) -> None:
+        entry = await self.getAuditLogs(guild)
+        if entry.target == user:
+            await doModlog(
+                self.bot,
+                guild,
+                entry.target,  # type: ignore
+                entry.user,
+                "unban",
+                entry.reason,
+            )
 
     @commands.Cog.listener("on_command_error")
     async def onCommandError(self, ctx, error) -> Optional[discord.Message]:
@@ -240,6 +266,7 @@ class EventHandler(commands.Cog, CogMixin):
             errors.MissingMuteRole,
             errors.CCommandNoPerm,
             errors.CCommandDisabled,
+            errors.NotNSFWChannel,
         )
 
         if isinstance(error, commands.CommandNotFound) or isinstance(
@@ -532,23 +559,37 @@ class EventHandler(commands.Cog, CogMixin):
             return
 
         with suppress(discord.Forbidden):
-            await asyncio.sleep(5)
-            entry = (
-                await guild.audit_logs(
-                    limit=1, action=discord.AuditLogAction.member_role_update
-                ).flatten()
-            )[0]
+            entry = await self.getAuditLogs(
+                guild, action=discord.AuditLogAction.member_role_update
+            )
 
-            if entry.target == member and entry.target._roles.has(mutedRole.id):
+            if entry.target == member and entry.target._roles.has(mutedRole.id):  # type: ignore
                 await doModlog(
                     self.bot,
                     member.guild,
-                    entry.target,
+                    entry.target,  # type: ignore
                     entry.user,
-                    "muted",
+                    "mute",
                     entry.reason,
                 )
 
     @commands.Cog.listener("on_member_unmuted")
     async def onMemberUnmuted(self, member: discord.Member, mutedRole: discord.Role):
-        return
+        if not (guild := member.guild):
+            # impossible to happened, but sure
+            return
+
+        with suppress(discord.Forbidden):
+            entry = await self.getAuditLogs(
+                guild, action=discord.AuditLogAction.member_role_update
+            )
+
+            if entry.target == member and not entry.target._roles.has(mutedRole.id):  # type: ignore
+                await doModlog(
+                    self.bot,
+                    member.guild,
+                    entry.target,  # type: ignore
+                    entry.user,
+                    "unmute",
+                    entry.reason,
+                )